% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style}
\alias{style}
\title{Style either the active file or active selection with tdc_style()}
\usage{
style(expand_args = getOption("tdcstyle.expand_args", FALSE))
}
\arguments{
\item{expand_args}{expand arguments for function calls to new lines if one is
on a new line}
}
\description{
Preference is given to the active selection if it exists.
}
