% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc_style.R
\name{tdc_style}
\alias{tdc_style}
\title{tdcstyle style guide}
\usage{
tdc_style()
}
\description{
This function provideds the style guide which you can supply to \code{{styler}} e.g. \verb{styler::style_file(<path>, style = tdcstyle::tdc_style)}
}
\details{
If your editor supports the {rstudioapi} you may have an easier time making a key binding for \code{\link[=style]{style()}}.
}
